More changes and tests

+++ tests/coreDox.Core.Tests/Projects/Config/ProjectConfigTests.cs
@@ -1,3 +1,4 @@
+using coreDox.Core.Exceptions;
 using coreDox.Core.Project;
 using coreDox.Core.Project.Config;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
@@ -14,22 +15,50 @@ namespace coreDox.Core.Tests.Projects
         [TestCleanup]
         public void TestCleanUp()
         {
-            Directory.Delete(_tmpPath, true);
+            if(Directory.Exists(_tmpPath))
+                Directory.Delete(_tmpPath, true);
+        }
+
+        [TestMethod]
+        public void ShouldCreateDefaultConfigSuccessfully()
+        {
+            //Arrange
+            Directory.CreateDirectory(_tmpPath);
+            var projectConfig = new DoxProjectConfig(_pluginRegistry, Path.Combine(_tmpPath, DoxProject.ConfigFileName));
+
+            //Act
+            projectConfig.CreateDefaultConfig();
+
+            //Assert
+            Assert.IsTrue(projectConfig.Exists);
         }
 
         [TestMethod]
         public void ShouldGetDoxConfigSuccessfully()
         {
             //Arrange
+            Directory.CreateDirectory(_tmpPath);
             var projectConfig = new DoxProjectConfig(_pluginRegistry, Path.Combine(_tmpPath, DoxProject.ConfigFileName));
-            var project = new DoxProject(projectConfig);
-            project.Load();
+            projectConfig.CreateDefaultConfig();
 
             //Act
-            var doxConfig = project.Config.GetConfigSection<DoxConfigSection>();
+            var doxConfig = projectConfig.GetConfigSection<DoxConfigSection>();
 
             //Assert
             Assert.IsNotNull(doxConfig);
+        }
+
+        [TestMethod]
+        [ExpectedException(typeof(CoreDoxException))]
+        public void ShouldThrowIfConfigFileNotPresentSuccessfully()
+        {
+            //Arrange
+            var projectConfig = new DoxProjectConfig(_pluginRegistry, Path.Combine(_tmpPath, DoxProject.ConfigFileName));
+
+            //Act
+            var doxConfig = projectConfig.GetConfigSection<DoxConfigSection>();
+
+            //Assert - Expects Exception
         }
     }
 }
