+++ tests/coreDox.Core.Tests/Projects/ProjectTests.cs
@@ -2,6 +2,7 @@ using coreDox.Core.Project;
 using coreDox.Core.Project.Config;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.IO;
+using System.Linq;
 
 namespace coreDox.Core.Tests.Projects
 {
@@ -13,11 +14,11 @@ namespace coreDox.Core.Tests.Projects
         [TestCleanup]
         public void TestCleanUp()
         {
-            Directory.Delete(_tmpPath, true);
+            if(Directory.Exists(_tmpPath)) Directory.Delete(_tmpPath, true);
         }
 
         [TestMethod]
-        public void ShouldCreateNewDefaultProjectOnLoadSuccessfully()
+        public void ShouldCreateDefaultProjectSuccessfully()
         {
             //Arrange
             var pluginRegistry = new PluginRegistry();
@@ -25,15 +26,28 @@ namespace coreDox.Core.Tests.Projects
             var project = new DoxProject(projectConfig);
 
             //Act
-            var loadResult = project.Load();
+            project.CreateMissingElements();
 
             //Assert
-            Assert.IsTrue(loadResult.AllSucceeded());
-
             Assert.IsTrue(Directory.Exists(Path.Combine(_tmpPath, DoxProject.AssetFolderName)));
             Assert.IsTrue(Directory.Exists(Path.Combine(_tmpPath, DoxProject.LayoutFolderName)));
             Assert.IsTrue(Directory.Exists(Path.Combine(_tmpPath, DoxProject.PagesFolderName)));
             Assert.IsTrue(File.Exists(Path.Combine(_tmpPath, DoxProject.ConfigFileName)));
+        }
+
+        [TestMethod]
+        public void ShouldReportErrorsOnNotValidProjectSuccessfully()
+        {
+            //Arrange
+            var pluginRegistry = new PluginRegistry();
+            var projectConfig = new DoxProjectConfig(pluginRegistry, Path.Combine(_tmpPath, DoxProject.ConfigFileName));
+            var project = new DoxProject(projectConfig);
+
+            //Act
+            var validationResult = project.IsProjectValid();
+
+            //Assert
+            Assert.IsTrue(validationResult.All(v => !v.Valid));
         }
     }
 }
